Fixing config according to what I understand
The config argument is not parsed as specified by the -h option.
I removed the call to CmdLineAppend library and instead parse the thing
locally.
I am not sure what extra processing the previous implementation was
doing (it seemed like it was appending extra things from the file but
only one path to the include path).
The other issue was the call to llvm::sys::path that was dropping the
basename regardless of whether or not it was a directory.
This yielded issues if you included a directory.

I think I misunderstand the usage of the config option so we might have
to reverse that change at some point.