Trying to specialize the template
Working on the specialization of the function template.
The normal method that is supposed to allow this has been declared as
protected, thus I have to rely on a weird trick that defines an
inherited class and statically cast my object to that type to access the
function.

Right now, the main challenge is to define a correct specializationInfo
to create a new version of the template.