NO_PREEMT macro added to the runtime
Tne NO_PREEMPT macro creates a local variable that increases a thread
local counter thus preventing preemption until the variable goes out of
scope.
For the moment, I automated the introduction of NO_PREEMPT and was
conservative, puting it even in empty methods.
We should discuss which ones are truly useful.

Also, we should wrap any snmalloc usage with a NO_PREEMPT.