Trying to fix implementation
Apparently still not working even after setting the context rather than
returning from the signal handler + not using rax as part of the return
path.

Next:
Not sure how to fix that. Maybe it has something to do with tls?
I might try a different implementation that solely relies on setcontext.